Chatbot Backend copy/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
pytest
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
